+++ package.json
@@ -16,7 +16,8 @@
     "oh-my-live2d": "^0.19.3",
     "ollama": "^0.5.11",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "zustand": "^5.0.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
