+++ src/components/Config.tsx
@@ -1,0 +1,8 @@
+export function Config() {
+
+  return (
+    <main>
+      Config
+    </main>
+  )
+}
