+++ src/components/History.tsx
@@ -1,0 +1,8 @@
+export function History() {
+
+  return (
+    <main>
+      History
+    </main>
+  )
+}
