+++ src/lib/api.browser.ts
@@ -1,6 +1,4 @@
-import type { SpeakApi, ChatApi, LoadLive2d, LoadChat, SaveChat, DeleteChat, API, Chat } from './types.ts'
-import ollama from 'ollama/browser'
-import { loadOml2d } from 'oh-my-live2d'
+import type { SpeakApi, LoadChat, SaveChat, DeleteChat, Chat } from './types.ts'
 import { get, update } from 'idb-keyval'
 
 let voices: SpeechSynthesisVoice[] = []
@@ -9,7 +7,7 @@ while (voices.length === 0) {
   voices = speechSynthesis.getVoices()
 }
 
-const speak: SpeakApi = (text: string) => {
+export const speak: SpeakApi = (text: string) => {
   const utterance = new SpeechSynthesisUtterance(text)
   const voice = voices.find(v => v.lang === 'zh-CN')
   if (voice) {
@@ -25,72 +23,13 @@ const speak: SpeakApi = (text: string) => {
   })
 }
 
-const model = 'qwen2.5:7b'
-const prompt = '你是一个心理咨询师, 名叫小叶子. 请你以支持的, 非指导性的方式陪伴对方, 帮助对方探索自己, 并在需要时提供帮助. 请不要回复长的和正式的内容, 避免说教, 表现得像一个真实、专业、共情的心理咨询师. 再次提醒: 回复务必要简短!'
-
-const chat: ChatApi = (messages) => {
-  return Promise.resolve(_chat(messages))
-}
-const _chat = async function* (messages: { role: string, content: string }[]) {
-  const response = await ollama.chat({
-    model,
-    messages: [{ role: 'system', content: prompt }, ...messages],
-    stream: true,
-  })
-  for await (const chunk of response) {
-    if (chunk.done) {
-      yield { response: chunk.message.content ?? '', done: true }
-    } else {
-      yield { response: chunk.message.content ?? '', done: false }
-    }
-  }
-}
-
-const loadLive2d: LoadLive2d = (element) => {
-  const live2d = loadOml2d({
-    parentElement: element,
-    dockedPosition: 'right',
-    menus: { disable: true },
-    sayHello: false,
-    tips: {
-      style: {
-        minWidth: '200px',
-      }
-    },
-    models: [
-      {
-        path: 'https://model.oml2d.com/cat-white/model.json',
-        position: [0, 20],
-      }
-    ],
-  })
-  return {
-    stop: () => live2d.clearTips(),
-    say: (text: string) => live2d.tipsMessage(text, 10000, Date.now()),
-    remove: () => { 
-      element.innerHTML = '' 
-    },
-  }
-}
-
-const loadChat: LoadChat = async () => {
+export const loadChat: LoadChat = async () => {
   const chats = await get<Chat[]>('chats')
   return chats ?? []
 }
-const saveChat: SaveChat = async (chat) => {
+export const saveChat: SaveChat = async (chat) => {
   await update<Chat[]>('chats', (chats) => [chat, ...(chats ?? [])])
 }
-const deleteChat: DeleteChat = async (uuid) => {
+export const deleteChat: DeleteChat = async (uuid) => {
   await update<Chat[]>('chats', (chats) => (chats ?? []).filter(chat => chat.uuid !== uuid))
 }
-
-const api: API = {
-  speak,
-  chat,
-  loadLive2d,
-  loadChat,
-  saveChat,
-  deleteChat,
-}
-
-export default api
