--- src/lib/config.ts
@@ -1,4 +1,0 @@
-import type { API } from './types.ts'
-import browserAPI from './api.browser.ts'
-
-export const api: API = browserAPI
